+++ packages/web3-eth-ens/src/ens.ts
@@ -176,8 +176,8 @@ export class ENS extends Web3Context<EthExecutionAPI & Web3NetAPI> {
 	 * @returns - The value content stored in the resolver for the specified key
 	 */
 	public async getText(ENSNameOrAddr: string | Address, key: string): Promise<string> {
-		if(isAddress(ENSNameOrAddr))
-			return this._resolver.getText(await(this._resolver.getName(ENSNameOrAddr,false)), key);
+		if (isAddress(ENSNameOrAddr))
+			return this._resolver.getText(await this._resolver.getName(ENSNameOrAddr, false), key);
 		return this._resolver.getText(ENSNameOrAddr, key);
 	}