+++ packages/web3-rpc-providers/src/types.ts
@@ -16,29 +16,29 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 export enum Transport {
-    HTTPS = "https",
-    WebSocket = "wss"
-};
+	HTTPS = 'https',
+	WebSocket = 'wss',
+}
 
 export enum Network {
-    ETH_MAINNET = "eth_mainnet",
-    ETH_GOERLI = "eth_goerli",
-    ETH_SEPOLIA = "eth_sepolia",
-    ETH_HOLESKY = "eth_holesky",
+	ETH_MAINNET = 'eth_mainnet',
+	ETH_GOERLI = 'eth_goerli',
+	ETH_SEPOLIA = 'eth_sepolia',
+	ETH_HOLESKY = 'eth_holesky',
 
-    POLYGON_MAINNET= "polygon_mainnet",
-    POLYGON_MUMBAI= "polygon_mumbai",
-    POLYGON_AMONY= "polygon_amony",
+	POLYGON_MAINNET = 'polygon_mainnet',
+	POLYGON_MUMBAI = 'polygon_mumbai',
+	POLYGON_AMONY = 'polygon_amony',
 
-    ARBITRUM_MAINNET = "arbitrum_mainnet",
-    ARBITRUM_SEPOLIA = "arbitrum_sepolia",
+	ARBITRUM_MAINNET = 'arbitrum_mainnet',
+	ARBITRUM_SEPOLIA = 'arbitrum_sepolia',
 
-    BASE_MAINNET = "base_mainnet",
-    BASE_SEPOLIA = "base_sepolia",
+	BASE_MAINNET = 'base_mainnet',
+	BASE_SEPOLIA = 'base_sepolia',
 
-    OPTIMISM_MAINNET = "optimism_mainnet",
-    OPTIMISM_SEPOLIA = "optimism_sepolia",
+	OPTIMISM_MAINNET = 'optimism_mainnet',
+	OPTIMISM_SEPOLIA = 'optimism_sepolia',
 
-    BNB_MAINNET = "bnb_mainnet",
-    BNB_TESTNET = "bnb_testnet"
-};
+	BNB_MAINNET = 'bnb_mainnet',
+	BNB_TESTNET = 'bnb_testnet',
+}