+++ packages/web3-eth-accounts/CHANGELOG.md
@@ -168,10 +168,13 @@ Documentation:
 
 - baseTransaction method updated (#7095)
 
-## [Unreleased]
+## [4.2.0]
+
 ### Added
 
 -   Added public function `signMessageWithPrivateKey` (#7174)
 
 ### Fixed
 -   Fix `TransactionFactory.registerTransactionType` not working, if there is a version mistatch between `web3-eth` and `web3-eth-accounts` by saving `extraTxTypes` at `globals`.  (#7197)
+
+## [Unreleased]